requirements.txt
alembic==1.13.1
fastapi==0.111.0
fastapi-mail==1.4.1
motor==3.4.0
passlib==1.7.4
psycopg2-binary==2.9.9
python-jose==3.3.0
python-socketio==5.11.2
SQLAlchemy==2.0.30